Refactor contexts to abstract `use` macro expansion logic
